#190 Move GlobalToArgument tests to a separate testing file
